+++ src/model/project.js
@@ -1,0 +1,9 @@
+class Project {
+  todoList = [];
+
+  constructor(pName) {
+    this.pName = pName;
+  }
+};
+
+export default Project;
