+++ src/model/todo.js
@@ -1,0 +1,17 @@
+class Todo {
+
+  constructor(title, description, dueDate, priority, project) {
+    this.title = title;
+    this.description = description;
+    this.dueDate = dueDate;
+    this.priority = priority;
+    this.done = false;
+    this.project = project;
+  }
+
+  markAsDone() {
+    this.done = true;
+  }
+}
+
+export default Todo;
